+++ .github/workflows/ci.yml
@@ -28,10 +28,10 @@ jobs:
         include:
           - os: macOS-latest
             url: https://github.com/graalvm/graalvm-ce-builds/releases/download/vm-20.0.0/graalvm-ce-java11-darwin-amd64-20.0.0.tar.gz
-            artifact: scalafmt-macos
+            artifact: git-mini-x86_64-apple-darwin
           - os: ubuntu-latest
             url: https://github.com/graalvm/graalvm-ce-builds/releases/download/vm-20.0.0/graalvm-ce-java11-linux-amd64-20.0.0.tar.gz
-            artifact: scalafmt-linux
+            artifact: git-mini-x86_64-pc-linux
     steps:
       - uses: actions/checkout@v2
       - uses: olafurpg/setup-scala@v6
@@ -40,8 +40,8 @@ jobs:
           java-version: graalvm@20.0.0
       - run: gu install native-image
       - run: sbt native-image
-      - run: cp gitflip/target/graalvm-native-image/gitflip git-flip
+      - run: cp gitmini/target/graalvm-native-image/gitmini git-mini
       - uses: actions/upload-artifact@master
         with:
           name: ${{ matrix.artifact }}
-          path: git-flip
+          path: git-mini
