--- gitflip/src/main/scala/gitflip/AmendCommand.scala
@@ -1,40 +1,0 @@
-package gitflip
-
-import metaconfig.cli.CliApp
-import metaconfig.cli.Command
-import GitflipEnrichments._
-import java.io.PrintWriter
-import java.nio.file.Path
-import scala.util.control.NonFatal
-import org.typelevel.paiges.Doc
-
-object AmendCommand extends Command[AmendOptions]("amend") {
-  override def description: Doc =
-    Doc.text("Edit the list of directories that are tracked by this mini-repo")
-  def run(value: Value, cli: CliApp): Int = {
-    val app = new Flip(cli)
-    app.withMinirepo("amend", value.minirepo) { name =>
-      if (app.megarepo == app.minirepo(name)) {
-        app.cli.error(
-          "can't amend the mega-krepo. To run amend, you must be inside a mini-repo."
-        )
-        1
-      } else {
-        Option(System.getenv("EDITOR")) match {
-          case None =>
-            app.cli.error(
-              "can't amend since the environment variable $EDITOR is not defined.\n\t" +
-                s"To fix this problem run: EDITOR=vim ${app.cli.arguments.mkString(" ")}"
-            )
-            1
-          case Some(editor) =>
-            app.execTty(s"$editor ${app.includes(name)}").ifSuccessful {
-              StartCommand.writeExclude(app, name)
-              app.commitAll(s"git-flip: amend mini-repo")
-            }
-        }
-      }
-    }
-  }
-
-}
