--- gitflip/src/main/scala/gitflip/StartOptions.scala
@@ -1,18 +1,0 @@
-package gitflip
-
-import java.nio.file.Path
-import metaconfig.annotation.ExtraName
-
-final case class StartOptions(
-    name: String = "",
-    @ExtraName("remainingArgs")
-    directories: List[Path] = Nil
-)
-
-object StartOptions {
-  val default = StartOptions()
-  implicit lazy val surface =
-    metaconfig.generic.deriveSurface[StartOptions]
-  implicit lazy val codec =
-    metaconfig.generic.deriveCodec[StartOptions](default)
-}
