--- gitflip/src/main/scala/gitflip/SyncOptions.scala
@@ -1,12 +1,0 @@
-package gitflip
-
-final case class SyncOptions(
-    rebase: Boolean = false
-)
-object SyncOptions {
-  val default = SyncOptions()
-  implicit lazy val surface =
-    metaconfig.generic.deriveSurface[SyncOptions]
-  implicit lazy val codec =
-    metaconfig.generic.deriveCodec[SyncOptions](default)
-}
