--- gitflip/src/test/scala/gitflip/CreateCommandSuite.scala
@@ -1,30 +1,0 @@
-package gitflip
-
-import java.io.File
-import scala.io.StdIn
-
-class CreateCommandSuite extends BaseSuite {
-  val hello = "hello/src/Hello.scala"
-  val goodbye = "goodbye/src/Goodbye.scala"
-  test("install") {
-    init(
-      s"""
-         |/$hello
-         |package hello
-         |object Hello
-         |/$goodbye
-         |package goodbye
-         |object Goodbye
-         |""".stripMargin
-    )
-    run("start", "--name", "hello", "hello")
-    assertEquals(lsFiles(), List(hello))
-    assertEquals(
-      execString("git", "diff", "origin/master"),
-      ""
-    )
-    run("uninstall")
-    assertEquals(lsFiles(), List(goodbye, hello))
-  }
-
-}
