+++ gitmini/src/main/scala/gitmini/SyncOptions.scala
@@ -1,0 +1,12 @@
+package gitmini
+
+final case class SyncOptions(
+    rebase: Boolean = false
+)
+object SyncOptions {
+  val default = SyncOptions()
+  implicit lazy val surface =
+    metaconfig.generic.deriveSurface[SyncOptions]
+  implicit lazy val codec =
+    metaconfig.generic.deriveCodec[SyncOptions](default)
+}
